+++ mac-linux/new-project
@@ -1,3 +1,3 @@
 #!/bin/bash
 pip3 install PyGithub 2>&1
-python3 ~/ProjectAutomation/automate-project.py
+python3 ~/ProjectAutomation/automate_project.py
