+++ windows/new-project.bat
@@ -1,4 +1,4 @@
 @ECHO OFF
 cd /D %~dp0..
 pip install PyGithub >NUL
-python automate-project.py
+python automate_project.py
